rename chat service to chats service

+++ Services/ChatsService.cs
@@ -1,3 +1,4 @@
+using System.Globalization;
 using Amazon.DynamoDBv2;
 using Amazon.DynamoDBv2.DataModel;
 using Amazon.DynamoDBv2.Model;
@@ -8,13 +9,14 @@ using static AnthemAPI.Common.Constants;
 
 namespace AnthemAPI.Services;
 
-public class ChatService
+public class ChatsService
 {
     private readonly IAmazonDynamoDB _client;
     private readonly DynamoDBContext _context;
+    private const int PAGE_LIMIT = 20;
     private const string TABLE_NAME = "Chats";
     
-    public ChatService(IAmazonDynamoDB client)
+    public ChatsService(IAmazonDynamoDB client)
     {
         _client = client;
         _context = new DynamoDBContext(client);
@@ -29,7 +31,7 @@ public class ChatService
         }
         catch (Exception e)
         {
-            return ServiceResult<Chat?>.Failure(e, $"Failed to load for {chatId}.", "ChatService.Load()");
+            return ServiceResult<Chat?>.Failure(e, $"Failed to load for {chatId}.", "ChatsService.Load()");
         }
     }
 
@@ -42,25 +44,30 @@ public class ChatService
         }
         catch (Exception e)
         {
-            return ServiceResult<Chat>.Failure(e, $"Failed to save for {chat.Id}.", "ChatService.Save()");
+            return ServiceResult<Chat>.Failure(e, $"Failed to save for {chat.Id}.", "ChatsService.Save()");
         }
     }
 
-    public async Task<ServiceResult<bool>> Delete(string chatId)
+    public async Task<ServiceResult<Chat?>> Delete(string chatId)
     {
         try
         {
             var load = await Load(chatId);
+
+            if (load.IsFailure || load.Data is null)
+                return load;
+            
             await _context.DeleteAsync(load.Data);
-            return ServiceResult<bool>.Success(true);
+
+            return ServiceResult<Chat?>.Success(load.Data);
         }
         catch (Exception e)
         {
-            return ServiceResult<bool>.Failure(e, $"Failed to delete {chatId}.", "ChatService.Delete()");
+            return ServiceResult<Chat?>.Failure(e, $"Failed to delete {chatId}.", "ChatsService.Delete()");
         }
     }
 
-    public async Task<ServiceResult<bool>> Update(string chatId, DateTime lastMessageAt, string preview)
+    public async Task<ServiceResult<Chat>> Update(string chatId, DateTime lastMessageAt, string preview)
     {
         try
         {
@@ -76,18 +83,28 @@ public class ChatService
                 {
                     [":preview"] = new AttributeValue { S = preview },
                     [":lastMessageAt"] = new AttributeValue { S = lastMessageAt.ToString("o") }
-                }
+                },
+                ReturnValues = ReturnValue.UPDATED_NEW
             };
 
             var response = await _client.UpdateItemAsync(update);
 
-            Console.WriteLine("attributes: " + response.Attributes);
+            var chat = new Chat
+            {
+                Id = response.Attributes["Id"].S,
+                Name = response.Attributes["Name"].S,
+                UserIds = response.Attributes["UserIds"].SS.ToHashSet(),
+                LastMessageAt = Helpers.ToDateTimeUTC(response.Attributes["LastMessageAt"].S),
+                Preview = response.Attributes["Preview"].S,
+                CreatorUserId = response.Attributes["CreatorUserId"].S,
+                CreatedAt = Helpers.ToDateTimeUTC(response.Attributes["CreatedAt"].S)
+            };
 
-            return ServiceResult<bool>.Success(true);
+            return ServiceResult<Chat>.Success(chat);
         }
         catch (Exception e)
         {
-            return ServiceResult<bool>.Failure(e, $"Failed to update for {chatId}.", "ChatService.Update()");
+            return ServiceResult<Chat>.Failure(e, $"Failed to update for {chatId}.", "ChatsService.Update()");
         }
     }
 
@@ -97,9 +114,9 @@ public class ChatService
         {
             var batches = new List<BatchGet<Chat>>();
 
-            for (int i = 0; i < chatIds.Count; i += DYNAMO_DB_BATCH_GET_ITEM_LIMIT)
+            for (int i = 0; i < chatIds.Count; i += DYNAMO_DB_BATCH_GET_LIMIT)
             {
-                List<string> ids  = chatIds.Skip(i).Take(DYNAMO_DB_BATCH_GET_ITEM_LIMIT).ToList();
+                List<string> ids  = chatIds.Skip(i).Take(DYNAMO_DB_BATCH_GET_LIMIT).ToList();
                 var batch = _context.CreateBatchGet<Chat>();
                 ids.ForEach(batch.AddKey);
                 batches.Add(batch);
@@ -110,15 +127,15 @@ public class ChatService
             var chats = batches
                 .SelectMany(b => b.Results)
                 .OrderByDescending(c => c.LastMessageAt)
-                .Skip(page > 1 ? Helpers.CalculatePaginationToken(page, CHAT_BATCH_LIMIT) : 0)
-                .Take(CHAT_BATCH_LIMIT)
+                .Skip(page > 1 ? (page - 1) * PAGE_LIMIT : 0)
+                .Take(PAGE_LIMIT)
                 .ToList();
 
             return ServiceResult<List<Chat>>.Success(chats);
         }
         catch (Exception e)
         {
-            return ServiceResult<List<Chat>>.Failure(e, "Failed to get all.", "ChatService.GetAll()");
+            return ServiceResult<List<Chat>>.Failure(e, "Failed to get page.", "ChatsService.GetPage()");
         }
     }
 }
